Update README.md
Added the link to Connexion, a Swagger/OpenAPI First framework for Python on top of Flask with automatic endpoint validation and OAuth2 support. This project is awesome because it allows you to write a Swagger specification, then maps the endpoints to your Python functions; many tools operate in reverse. You can describe your REST API in as much detail as you want; then Connexion guarantees that it will work as you specified. The documentation is thorough and up to date, and there are examples included. It's very much in active development and generating queries and contributions from the community; we would like to involve even more external contributors. :)